+++ Steevanb/Sniffs/Arrays/DisallowShortArraySyntaxSpacesSniff.php
@@ -2,8 +2,15 @@
 
 declare(strict_types=1);
 
+namespace steevanb\PhpCodeSniffs\Steevanb\Sniffs\Arrays;
+
+use PHP_CodeSniffer\{
+    Files\File,
+    Sniffs\Sniff
+};
+
 /** Disallow spaces after [ and before ] */
-class Steevanb_Sniffs_Arrays_DisallowShortArraySyntaxSpacesSniff implements PHP_CodeSniffer_Sniff
+class DisallowShortArraySyntaxSpacesSniff implements Sniff
 {
     public function register(): array
     {
@@ -11,7 +18,7 @@ class Steevanb_Sniffs_Arrays_DisallowShortArraySyntaxSpacesSniff implements PHP_
     }
 
     /** @param int $stackPtr */
-    public function process(PHP_CodeSniffer_File $phpcsFile, $stackPtr): void
+    public function process(File $phpcsFile, $stackPtr): void
     {
         $token = $phpcsFile->getTokens()[$stackPtr];
         if ($token['code'] === 'PHPCS_T_OPEN_SHORT_ARRAY') {
@@ -22,23 +29,31 @@ class Steevanb_Sniffs_Arrays_DisallowShortArraySyntaxSpacesSniff implements PHP_
     }
 
     /** @param int $stackPtr */
-    protected function processOpenShortArray(PHP_CodeSniffer_File $phpcsFile, $stackPtr): self
+    protected function processOpenShortArray(File $phpcsFile, $stackPtr): self
     {
         $nextToken = $phpcsFile->getTokens()[$stackPtr + 1];
         if ($nextToken['code'] === T_WHITESPACE && substr($nextToken['content'], 0, 1) !== "\n") {
-            $phpcsFile->addErrorOnLine('Short array syntax should not begin with spaces', $nextToken['line']);
+            $phpcsFile->addErrorOnLine(
+                'Short array syntax should not begin with spaces',
+                $nextToken['line'],
+                'NoSpaceAtOpen'
+            );
         }
 
         return $this;
     }
 
     /** @param int $stackPtr */
-    protected function processCloseShortArray(PHP_CodeSniffer_File $phpcsFile, $stackPtr): self
+    protected function processCloseShortArray(File $phpcsFile, $stackPtr): self
     {
         $previousToken = $phpcsFile->getTokens()[$stackPtr - 1];
         $previousToken2 = $phpcsFile->getTokens()[$stackPtr - 2];
         if ($previousToken['code'] === T_WHITESPACE && $previousToken2['code'] !== T_WHITESPACE) {
-            $phpcsFile->addErrorOnLine('Short array syntax should not end with spaces', $previousToken['line']);
+            $phpcsFile->addErrorOnLine(
+                'Short array syntax should not end with spaces',
+                $previousToken['line'],
+                'NoSpaceAtClose'
+            );
         }
 
         return $this;