+++ changelog.md
@@ -1,11 +1,18 @@
+### [2.0.0](../../compare/1.0.2...2.0.0) - 2019-04-16
+
+- Update `squizlabs/php_codesniffer` to 3.2.*.
+- Fix `DisallowMultipleReturnSniff` when a method contains a closure.
+- Fix `DisallowSelfSniff` when used as method return type.
+- Add color to sniff name when `-s` parameter is used with `./vendor/bin/phpcs`.
+
 ### [1.0.2](../../compare/1.0.1...1.0.2) - 2019-01-29
 
-- Fix `Steevanb_Sniffs_Functions_DisallowMultipleReturnSniff::addAllowedFunction()`
+- Fix `Steevanb_Sniffs_Functions_DisallowMultipleReturnSniff::addAllowedFunction()`.
 
 ### [1.0.1](../../compare/1.0.0...1.0.1) - 2019-01-16
 
-- Fix `Steevanb_Sniffs_Uses_GroupUsesSniff::addThirdLevelPrefix()`
+- Fix `Steevanb_Sniffs_Uses_GroupUsesSniff::addThirdLevelPrefix()`.
 
 ### 1.0.0 - 2019-01-16
 
-- First version
+- First version.